dev-Product-Service

+++ server/src/main/java/papa_hant/first/server/routes/typeProduct/service/TypeProductService.java
@@ -1,5 +1,6 @@
 package papa_hant.first.server.routes.typeProduct.service;
 
+import ch.qos.logback.core.encoder.EchoEncoder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -8,6 +9,7 @@ import papa_hant.first.server.App.models.Response.Types.ResponseData;
 import papa_hant.first.server.routes.typeProduct.entity.TypeProduct;
 import papa_hant.first.server.routes.typeProduct.entity.dto.TypeProductDto;
 import papa_hant.first.server.routes.typeProduct.repository.TypeProductRepository;
+import papa_hant.first.server.routes.user.service.UserService;
 
 /**
  * ## Сервис типов товара
@@ -15,19 +17,23 @@ import papa_hant.first.server.routes.typeProduct.repository.TypeProductRepositor
  * @author Горелов Дмитрий
  * */
 
-// TODO: Добавить проверку на админа
-
 @Service
 @RequiredArgsConstructor
 public class TypeProductService {
     private final TypeProductRepository repository;
 
+    private final UserService userService;
+
     /**
      * Добавить новый Тип
      * */
 
-    public Response saveType(TypeProductDto dto){
+    public Response saveType(TypeProductDto dto, String token){
         try {
+            if (!userService.roleVerification(token)){
+                return new Response(HttpStatus.FORBIDDEN.value(), "Нет доступа");
+            }
+
             if (this.repository.findByName(dto.getName()) != null){
                 return new Response(HttpStatus.CONTINUE.value(), "Товар с таким именем уже есть");
             }
@@ -46,8 +52,12 @@ public class TypeProductService {
      * Изменить существующий Тип
      * */
 
-    public Response updateType(Long id, TypeProductDto dto){
+    public Response updateType(Long id, TypeProductDto dto, String token){
         try {
+            if (!userService.roleVerification(token)){
+                return new Response(HttpStatus.FORBIDDEN.value(), "Нет доступа");
+            }
+
             TypeProduct type = this.repository.findById(id).get();
 
             if (type == null){
@@ -70,8 +80,12 @@ public class TypeProductService {
 
     // TODO : Добавить проверку, что не используется в товарах
 
-    public Response deleteById(Long id){
+    public Response deleteById(Long id, String token){
         try {
+            if (!userService.roleVerification(token)){
+                return new Response(HttpStatus.FORBIDDEN.value(), "Нет доступа");
+            }
+
             this.repository.deleteById(id);
             return new Response(HttpStatus.OK.value(), "Успешно удалено");
         }catch (Exception err){
@@ -83,8 +97,12 @@ public class TypeProductService {
      * Найти один Тип
      * */
 
-    public Response findById(Long id){
+    public Response findById(Long id, String token){
         try {
+            if (!userService.roleVerification(token)){
+                return new Response(HttpStatus.FORBIDDEN.value(), "Нет доступа");
+            }
+
             return new ResponseData<>(HttpStatus.OK.value(),
                     "Успешно получено",
                     this.repository.findById(id));
@@ -104,6 +122,15 @@ public class TypeProductService {
                     this.repository.findAll());
         }catch (Exception err){
             return new Response(HttpStatus.INTERNAL_SERVER_ERROR.value(), err.getMessage());
+        }
+    }
+
+    public TypeProduct findByIdNotAuth(Long id){
+        try {
+            return this.repository.findById(id).get();
+        }catch (Exception err){
+            System.out.println(err.getMessage());
+            return null;
         }
     }
 }
